// src/app/interfaces/current-weatherapi.model.ts
import { Weatherapi } from './weatherapi.model'
export interface CurrentWeatherapi {
    coord: {
        lon: number;
        lat: number;
    };
    weather: Weatherapi[];
    base: string;
    main: {
        temp: number;
        feels_like: number;
        temp_min: number;
        temp_max: number;
        pressure: number;
        humidity: number;
    };
    visibility: number;
    wind: {
        speed: number;
        deg: number;
    };
    clouds: {
        all: number;
    };
    dt: number;
    sys: {
        type: number;
        id: number;
        message: number;
        country: string;
        sunrise: number;
        sunset: number;
    };
    timezone: number;
    id: number;
    name: string;
    cod: number;
}